Retry on 5xx response from Vault, log status code
When Vault returned a non-200 status code, we would still treat the
response as a success response and try to parse it. This would lead us
to log an "[ERROR] Key 'xxx' not found" message, which is misleading.
This commit changes that, so we only do the parse on a 200 response, and
on a non-200 response we log the status code and response body.

Furthermore, add a retry. If Vault returns 500, 503, or 504, then we
retry the request at most 2 times, with a random delay in between.

This should make starting our services more resilient.