Use string instead of read lookup where needed
This fixes a small bug/ergonomics issue where strings passed via
environment variables needed to be in Haskell format. From now on,
we just use the plain string.

This means users can provide strings like so:

    VAULT_HOST="vault.company.com" vaultenv ...

Instead of having to provide them as a Haskell value:

    VAULT_HOST='"vault.company.con"' vaultenv ...

This isn't a backwards incompatible change, since the previous
behavior never made it into a release.