Build Aeson with Bazel
To fix this, I had to fix the cbits in the "primitive" package, for
which I had to have a GHC target with :threaded-rts target to get the
right includes. Also, to avoid a linking error, we need to remove a
module from the "hashable" package. In the Cabal file it was guarded,
used only for older versions of GHC. So remove it.