Fix cbits for a few libraries
There were missing c files, so even though the symbol was declared, it
was not defined, and it was loaded only at runtime, causing a failure in
the loader. Also include the implementations.

For the Cryptonite package the situation is a bit more complex as it
contains a ton of C code. I originally did not include all of it. I
included a bit more now, but not enough to make Vaultenv run.