Make Token an optional argument
Vault agents usually take care of authentication, so the authentication
header is usually not necessary. If the Token is not set, Vaultenv will
simply not send the header.

In case no agent was used, Vault will respond with 400/Bad Request and
the message 'missing client token'. 400 errors are not retried, so this
error is nearly as discoverable as it used to be.