Return a better error when trying to use non-string keys in secret
The approach in the previous commit would have caused a "key not found"
error if you tried to use a non-string key, as they were silently
dropped.

This commit changes VaultData to directly refer to an Aeson.Value and
instead does the string check in lookupSecrets, which has the benefit
that it simplifies the code a fair bit.

The parser has to change from doing `parseV1` first to doing `parseV2`
first, since a V2 response is now considered valid by the `parseV1`
parser.