Add auth method CLI options
This is to prepare for Kubernetes support. To authenticate with a
Kubernetes service account, we need to make one additional request that
then returns the X-Vault-Token. My plan is to check for AuthKubernetes
before we start retrieving secrets, and to switch it back to
AuthVaultToken once the token has been obtained, so most of the logic
does not need to change.

For now, this only adds the CLI options for this, but trying to use it
would crash.