Print the response body in case of invalid json
This is a bit of a trade-off between safety and debuggability. If the
response could not be parsed, but still contained a secret, this might
log secrets to stdout, and they might end up in places where secrets are
not supposed to go. But on the other hand, the error message that some
json failed to parse, is not very useful if you can't see what the json
what that we tried to parse, and if it fails, it is likely a programming
or configuration error, and the response will not contain secrets.