Rewrite the secrets file parser
There is a place for true parsers, but the Megaparsec one here was
fragile. I wanted to add support for comment lines to it but it was not
so clear where, and the current parser had this pretty dangerous failure
mode of ignoring half the input. Instead of trying to deal with that,
for Vaultenv's secrets file format, it is so simple, just splitting on
lines and then splitting on = and # should be fine.